Refactor: AdminWrite 템플릿 분리
