+++ src/components/organisms/index.ts
@@ -1,4 +1,3 @@
-export { default as AdminWrite } from './AdminWrite';
 export { default as CategoryList } from './CategoryList';
 export { default as ChallengeDetailCard } from './ChallengeDetailCard';
 export { default as ChallengeEditCard } from './ChallengeEditCard';
